sprint1
버그 수정
